refactor: move theme to own file
KBDEV-1500
